upgraded database to postgres, because of perfomance issues

+++ Script.sql
@@ -1,5 +1,5 @@
 CREATE table clientes (
-	cedula bigint(20) primary key,
+	cedula bigint primary key,
 	direccion varchar(40) not null,	
 	email varchar(40) not null,
 	nombre varchar(40) not null,
@@ -7,7 +7,7 @@ CREATE table clientes (
 );
 
 CREATE table usuarios (
-	cedula bigint(20) primary key,
+	cedula bigint primary key,
 	email varchar(40) not null,	
 	nombre varchar(40) not null,
 	password varchar(40) not null,
@@ -15,7 +15,7 @@ CREATE table usuarios (
 );
 
 CREATE table proveedores (
-	nit bigint(20) primary key,
+	nit bigint primary key,
 	ciudad varchar(40) not null,	
 	direccion varchar(40) not null,
 	nombre varchar(40) not null,
@@ -23,36 +23,34 @@ CREATE table proveedores (
 );
 
 CREATE table productos (
-	codigo bigint(20) primary key,
-	nit_proveedor bigint(20) not null,
-	iva_compra double not null,	
+	codigo bigint primary key,
+	nit_proveedor bigint not null,
+	iva_compra decimal not null,	
 	nombre_producto varchar(40) not null,
-	precio_compra double not null,
-	precio_venta double not null,
+	precio_compra decimal not null,
+	precio_venta decimal not null,
 	foreign key(nit_proveedor) references proveedores(nit)
 );
 
 CREATE table ventas (
-	codigo bigint(20) primary key,
-	cedula_cliente bigint(20) not null,
-	cedula_usuario bigint(20) not null,
-	iva_venta double not null,	
-	total_venta double not null,
-	valor_venta double not null,
+	codigo bigint primary key,
+	cedula_cliente bigint not null,
+	cedula_usuario bigint not null,
+	iva_venta decimal not null,	
+	total_venta decimal not null,
+	valor_venta decimal not null,
 	foreign key(cedula_cliente) references clientes(cedula),
 	foreign key(cedula_usuario) references usuarios(cedula)
 );
 
 CREATE table detalle_ventas (
-	codigo bigint(20) primary key,
-	codigo_producto bigint(20) not null,
-	codigo_venta bigint(20) not null,
-	cantidad_producto int(11) not null,	
-	valor_total double not null,
-	valor_venta double not null,
-	valor_iva double not null,
+	codigo bigint primary key,
+	codigo_producto bigint not null,
+	codigo_venta bigint not null,
+	cantidad_producto int not null,	
+	valor_total decimal not null,
+	valor_venta decimal not null,
+	valor_iva decimal not null,
 	foreign key(codigo_producto) references productos(codigo),
 	foreign key(codigo_venta) references ventas(codigo)
 );
-
-
